+++ api_server/src/main/java/com/jzy/api/service/home/impl/DealerHomeCateServiceImpl.java
@@ -35,7 +35,7 @@ public class DealerHomeCateServiceImpl extends GenericServiceImpl<HomeRecommendC
 
 
     @Override
-    public Integer listPageCount(DealerHomeCateListCnd listCnd) {
+    public Long listPageCount(DealerHomeCateListCnd listCnd) {
         return mapper.listPageCount(listCnd);
     }
